requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Flask==3.0.3
fonttools==4.54.1
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.1
matplotlib==3.9.2
mysql-connector==2.2.9
mysql-connector-python==9.0.0
numpy==2.1.2
packaging==24.1
pillow==10.4.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
six==1.16.0
Werkzeug==3.0.4
